Filmorate #9 Фикс для прохождения тестов режиссеров в коллекции add-search

+++ src/main/java/ru/yandex/practicum/filmorate/storage/RatingStorage.java
@@ -10,7 +10,7 @@ public interface RatingStorage {
 
     Collection<Rating> getAll();
 
-    Optional<Rating> getById(int id);
+    Optional<Rating> getById(long id);
 
     int getCountOfMpa();
